backend/pyproject.toml
[project]
name = "voliti"
version = "0.1.0"
description = "基于教练协议的长期行为对齐系统"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.0",
    "jinja2>=3.1.6",
    "langchain-openai>=0.3.0",
    "langgraph>=1.1.0",
    "python-dotenv>=1.2.1",
    "langgraph-cli>=0.4.18",
    "openai>=1.60.0",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.12",
    "pytest>=9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voliti"]

[tool.pytest.ini_options]
pythonpath = ["src"]
